+++ libtransmission/rpcimpl.cc
@@ -2445,9 +2445,9 @@ using SessionAccessors = std::pair<SessionGetter, SessionSetter>;
         [](tr_session const& src) -> tr_variant { return src.queueStalledMinutes(); },
         [](tr_session& tgt, tr_variant const& src, ErrorInfo& /*err*/)
         {
-            if (auto const val = src.value_if<int64_t>())
+            if (auto const val = src.value_if<size_t>())
             {
-                tr_sessionSetQueueStalledMinutes(&tgt, static_cast<int>(*val));
+                tr_sessionSetQueueStalledMinutes(&tgt, *val);
             }
         });
